+++ main.rkt
@@ -5,6 +5,7 @@
 (provide css-code
          c-code
          cpp-code
+         objc-code
          csv-code
          html-code
          js-code
@@ -22,6 +23,7 @@
          cssblock
          cblock
          cppblock
+         objcblock
          csvblock
          htmlblock
          jsblock
@@ -39,6 +41,7 @@
          cssblock0
          cblock0
          cppblock0
+         objcblock0
          csvblock0
          htmlblock0
          jsblock0
@@ -67,6 +70,7 @@
 (define-syntax-rule (css-code . rest) (api:css-code . rest))
 (define-syntax-rule (c-code . rest) (api:c-code . rest))
 (define-syntax-rule (cpp-code . rest) (api:cpp-code . rest))
+(define-syntax-rule (objc-code . rest) (api:objc-code . rest))
 (define-syntax-rule (csv-code . rest) (api:csv-code . rest))
 (define-syntax-rule (html-code . rest) (api:html-code . rest))
 (define-syntax-rule (js-code . rest) (api:js-code . rest))
@@ -84,6 +88,7 @@
 (define-syntax-rule (cssblock . rest) (api:cssblock . rest))
 (define-syntax-rule (cblock . rest) (api:cblock . rest))
 (define-syntax-rule (cppblock . rest) (api:cppblock . rest))
+(define-syntax-rule (objcblock . rest) (api:objcblock . rest))
 (define-syntax-rule (csvblock . rest) (api:csvblock . rest))
 (define-syntax-rule (htmlblock . rest) (api:htmlblock . rest))
 (define-syntax-rule (jsblock . rest) (api:jsblock . rest))
@@ -101,6 +106,7 @@
 (define-syntax-rule (cssblock0 . rest) (api:cssblock0 . rest))
 (define-syntax-rule (cblock0 . rest) (api:cblock0 . rest))
 (define-syntax-rule (cppblock0 . rest) (api:cppblock0 . rest))
+(define-syntax-rule (objcblock0 . rest) (api:objcblock0 . rest))
 (define-syntax-rule (csvblock0 . rest) (api:csvblock0 . rest))
 (define-syntax-rule (htmlblock0 . rest) (api:htmlblock0 . rest))
 (define-syntax-rule (jsblock0 . rest) (api:jsblock0 . rest))