+++ CHANGELOG.md
@@ -9,8 +9,9 @@
   - C: `c-code`, `cblock`, `cblock0`
   - C++: `cpp-code`, `cppblock`, `cppblock0`
   - CSV: `csv-code`, `csvblock`, `csvblock0`
-  - JSON: `json-code`, `jsonblock`, `jsonblock0`
+   - JSON: `json-code`, `jsonblock`, `jsonblock0`
   - Markdown: `markdown-code`, `markdownblock`, `markdownblock0`
+  - Objective-C: `objc-code`, `objcblock`, `objcblock0`
   - Racket: `racket-code`, `racketblock`, `racketblock0`
   - Rhombus: `rhombus-code`, `rhombusblock`, `rhombusblock0`
   - Swift: `swift-code`, `swiftblock`, `swiftblock0`